Fix $PS1, add i3 and i3blocks configs
bashrc:
Not enclosing escape sequences between \[ \] causes wrapping to be
weird. Wrapping escape sequences this way also respects the actual
whitespace in PS1 (meaning it won't take up space; more on that in Shott's
The Linux Command Line)

i3 and i3blocks:
I have decided that, for a bit of historical record (and since I've seen that
storage space isn't that big of a problem for me usually), I will keep all
dotfiles for every wm and status bar that I ever configure. This also has the
purpose of being able to bootstrap my installation process, since StumpWM is on
the AUR, and thus isn't as readily available as i3.