Trim and reorder config files
README:
Updated the readme to better reflect that the repo will constitute a sort of
digital garden with multiple rices

Xresources:
Changed the font of the terminal and removed comments

xinitrc:
Simply changed the wm, but this will eventually be overwritten by the
session-choose script

bashrc:
Trimmed the file down to the bare minimum, hopefully I'll expand on it

i3:
Trimmed the config and reorganized it into sections to try to better
reflect configuration options that may have stuff in common

picom:
Changed the opacity rule class to urxvt. Must add the same to alacritty
and check what other options are available

stumpwm:
Commented slynk code because I haven't installed that. The dotfile
unloading script should handle these kinds of situations where there are
dependencies among dotfiles.