Update ranger-cd binding/alias
So, guix weirdness, but basically just doing ". ranger"
doesn't work because SHLVL stays at 1; guix's ranger
(i.e. `which ranger`) sets up some python and w3m stuff
and calls exec -a "$0" ..., so quitting an invocation
of ranger-cd quits the whole shell! This commit adds the
bash script that is shown in the ranger python executable
and binds C-o to source that (explicitly btw, since having
shebangs in guix, which doesn't respect FHS is a whole 'nother
story........)